Moved CSDiffRunner logic to CSDiff
